Initial build and successful test of Stripe incorporation
