--- server/constants/stripe.js
@@ -1,10 +1,0 @@
-const configureStripe = require('stripe');
-
-const STRIPE_SECRET_KEY =
-  process.env.NODE_ENV === 'production'
-    ? process.env.STRIPE_LIVE_SECRET_KEY
-    : process.env.STRIPE_TEST_SECRET_KEY;
-
-const stripe = configureStripe(STRIPE_SECRET_KEY);
-
-module.exports = stripe;
