--- server/constants/stripeServerConfig.js
@@ -1,10 +1,0 @@
-const path = require('path');
-
-const SERVER_PORT = 8080;
-
-const SERVER_CONFIGS = {
-  PRODUCTION: process.env.NODE_ENV === 'production',
-  PORT: process.env.PORT || SERVER_PORT
-};
-
-module.exports = SERVER_CONFIGS;
